+++ e2e-tests/src/utils.rs
@@ -1,73 +1,16 @@
-use codec::Compact;
+pub mod fee;
+pub mod types;
+pub mod accounts;
+
 use common::create_connection;
 use log::info;
-use sp_core::{sr25519, Pair};
-use sp_runtime::{generic, traits::BlakeTwo256, AccountId32, MultiAddress};
-use substrate_api_client::rpc::WsRpcClient;
-use substrate_api_client::{AccountId, Api, Balance, UncheckedExtrinsicV4};
+use sp_core::Pair;
+use sp_runtime::AccountId32;
+use substrate_api_client::{AccountId, Balance, UncheckedExtrinsicV4};
 
 use crate::config::Config;
-
-pub type BlockNumber = u32;
-pub type Header = generic::Header<BlockNumber, BlakeTwo256>;
-pub type KeyPair = sr25519::Pair;
-pub type Connection = Api<KeyPair, WsRpcClient>;
-pub type TransferTransaction =
-    UncheckedExtrinsicV4<([u8; 2], MultiAddress<AccountId, ()>, Compact<u128>)>;
-
-pub fn keypair_from_string(seed: String) -> KeyPair {
-    KeyPair::from_string(&seed, None).expect("Can't create pair from seed value")
-}
-
-pub fn accounts(seeds: Option<Vec<String>>) -> Vec<KeyPair> {
-    let seeds = seeds.unwrap_or_else(|| {
-        vec![
-            "//Damian".into(),
-            "//Tomasz".into(),
-            "//Zbyszko".into(),
-            "//Hansu".into(),
-        ]
-    });
-    seeds.into_iter().map(keypair_from_string).collect()
-}
-
-pub fn get_sudo(config: Config) -> KeyPair {
-    match config.sudo {
-        Some(seed) => keypair_from_string(seed),
-        None => accounts(config.seeds)[0].to_owned(),
-    }
-}
-
-#[derive(Debug)]
-pub struct FeeInfo {
-    pub fee_without_weight: Balance,
-    pub unadjusted_weight: Balance,
-    pub adjusted_weight: Balance,
-}
-
-pub fn get_tx_fee_info(connection: &Connection, tx: &TransferTransaction) -> FeeInfo {
-    let unadjusted_weight = connection
-        .get_payment_info(&tx.hex_encode(), None)
-        .unwrap()
-        .unwrap()
-        .weight as Balance;
-
-    let fee = connection
-        .get_fee_details(&tx.hex_encode(), None)
-        .unwrap()
-        .unwrap();
-    let inclusion_fee = fee.inclusion_fee.unwrap();
-
-    FeeInfo {
-        fee_without_weight: inclusion_fee.base_fee + inclusion_fee.len_fee + fee.tip,
-        unadjusted_weight,
-        adjusted_weight: inclusion_fee.adjusted_weight_fee,
-    }
-}
-
-pub fn get_free_balance(account: &AccountId32, connection: &Connection) -> Balance {
-    connection.get_account_data(account).unwrap().unwrap().free
-}
+use crate::utils::types::{Connection, KeyPair, TransferTransaction};
+use crate::utils::accounts::accounts;
 
 pub fn setup_for_transfer(config: Config) -> (Connection, AccountId32, AccountId32) {
     let Config { node, seeds, .. } = config;
